Add Identifying characteristics to apperance form
