+++ src/containers/profile/edit/basicinformation/schemas/AppearanceSchemas.js
@@ -6,6 +6,7 @@ import {
   WEIGHT_FQN,
   EYE_COLOR_FQN,
   HAIR_COLOR_FQN,
+  DESCRIPTION_FQN,
 } from '../../../../../edm/DataModelFqns';
 import {
   EYE_COLOR_VALUES,
@@ -14,6 +15,7 @@ import {
 
 const { getEntityAddressKey, getPageSectionKey } = DataProcessingUtils;
 const {
+  IDENTIFYING_CHARACTERISTICS_FQN,
   PHYSICAL_APPEARANCE_FQN,
 } = APP_TYPES_FQNS;
 
@@ -43,6 +45,10 @@ const schema = {
           title: 'Hair Color',
           enum: HAIR_COLOR_VALUES
         },
+        [getEntityAddressKey(0, IDENTIFYING_CHARACTERISTICS_FQN, DESCRIPTION_FQN)]: {
+          type: 'string',
+          title: 'Scars/Marks/Tattoos'
+        }
       }
     }
   }
@@ -66,6 +72,9 @@ const uiSchema = {
     [getEntityAddressKey(0, PHYSICAL_APPEARANCE_FQN, HAIR_COLOR_FQN)]: {
       classNames: 'column-span-3'
     },
+    [getEntityAddressKey(0, IDENTIFYING_CHARACTERISTICS_FQN, DESCRIPTION_FQN)]: {
+      classNames: 'column-span-12'
+    }
   }
 };
